requirements.txt
crewai==0.119.0
langchain==0.3.25
langchain-google-genai==2.1.5
langchain-groq==0.3.2
langchain-openai==0.3.14
langgraph==0.3.31
python-telegram-bot==21.11.1
tavily-python==0.5.0
wikipedia==1.4.0
arxiv==2.1.3
mem0ai==0.1.101
chromadb==0.5.23
reportlab==4.2.5
python-dotenv==1.0.1
requests==2.32.3
openai==1.78.0
